fix: aggregate text & functionCall(s) correctly (#497)
* fix: aggregate text & functionCall(s) correctly

content parts should have either text or functionCall set, but never both. When aggregating, combine the text (as before), but add functionCall(s) as new parts. Then when done, if the text from part[0] is empty, that part can be removed.

fixes #494

* fix: aggregate text & functionCall(s) correctly

don't delete the text part, because it could just be added back again when working with multiple responses (and the end up as `"undefined\n"` for example). Leave the text part, and always push functions as new parts.

fixes #494

* fix: aggregate text & functionCall(s) correctly

the AI API gives an error for empty text, so the empty text has to be deleted after aggregating all responses

fixes #494

* chore: fix style using script

---------

Co-authored-by: Yvonne Yu <150068659+yyyu-google@users.noreply.github.com>